Constitute data_classes.py to fit issue #9
1. Merge data-related objects into a file.
2. Named it as data_classes.py
